python/requirements.txt
numpy>=1.24
scipy>=1.10
rasterio>=1.3
pyproj>=3.4
Pillow>=9.0
matplotlib>=3.6
numba>=0.57
